pyproject.toml
[project]
name = "aif"
version = "2.0.0"
description = "Analytical Intelligence Framework for developing a DWH and analytical applications."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dagster>=1.10.17",
    "dagster-cloud>=1.10.17",
    "dagster-webserver>=1.10.17",
    "dagstermill>=0.26.17",
    "ipython>=9.2.0",
    "jupyter>=1.1.1",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pandera>=0.24.0",
    "psycopg>=3.2.9",
    "psycopg-binary>=3.2.9",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "seaborn>=0.13.2",
    "yfinance>=0.2.61",
]

[tool.uv]
default-groups = "all"
